refactor(rename): rename file and component name to product table
